renaming PrintStyle component to PrintStyleSheet
